requirements.txt
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
requests==2.32.4
notion-client==2.0.0
aiohttp==3.12.14
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0